+++ core/clients/key_flow_test.go
@@ -178,8 +178,8 @@ func TestSetToken(t *testing.T) {
 					AccessToken:  accessToken,
 					ExpiresIn:    int(timestamp.Unix()),
 					RefreshToken: tt.refreshToken,
-					Scope:        defaultScope,
-					TokenType:    defaultTokenType,
+					Scope:        "",
+					TokenType:    "Bearer",
 				}
 				if !cmp.Equal(expectedKeyFlowToken, keyFlow.token) {
 					t.Errorf("The returned result is wrong. Expected %+v, got %+v", expectedKeyFlowToken, keyFlow.token)
@@ -194,25 +194,25 @@ func TestTokenExpired(t *testing.T) {
 	tests := []struct {
 		desc              string
 		tokenInvalid      bool
-		tokenExpiresAt    time.Time
+		tokenDuration     time.Duration
 		expectedErr       bool
 		expectedIsExpired bool
 	}{
 		{
 			desc:              "token valid",
-			tokenExpiresAt:    time.Now().Add(time.Hour),
+			tokenDuration:     time.Hour,
 			expectedErr:       false,
 			expectedIsExpired: false,
 		},
 		{
 			desc:              "token expired",
-			tokenExpiresAt:    time.Now().Add(-time.Hour),
+			tokenDuration:     -time.Hour,
 			expectedErr:       false,
 			expectedIsExpired: true,
 		},
 		{
 			desc:              "token almost expired",
-			tokenExpiresAt:    time.Now().Add(tokenExpirationLeeway),
+			tokenDuration:     tokenExpirationLeeway,
 			expectedErr:       false,
 			expectedIsExpired: true,
 		},
@@ -228,9 +228,7 @@ func TestTokenExpired(t *testing.T) {
 			var err error
 			token := "foo"
 			if !tt.tokenInvalid {
-				token, err = jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.RegisteredClaims{
-					ExpiresAt: jwt.NewNumericDate(tt.tokenExpiresAt),
-				}).SignedString([]byte("test"))
+				token, err = signToken(tt.tokenDuration)
 				if err != nil {
 					t.Fatalf("failed to create token: %v", err)
 				}
@@ -442,10 +440,9 @@ func TestKeyFlow_Do(t *testing.T) {
 							res.Header().Set("Content-Type", "application/json")
 
 							token := &TokenResponseBody{
-								AccessToken:  testBearerToken,
-								ExpiresIn:    2147483647,
-								RefreshToken: testBearerToken,
-								TokenType:    "Bearer",
+								AccessToken: testBearerToken,
+								ExpiresIn:   2147483647,
+								TokenType:   "Bearer",
 							}
 
 							if err := json.NewEncoder(res.Body).Encode(token); err != nil {